fix: move code to the right folder and delete unnecessary comments
